Move TaggedEventDistanceComparator into eventseries package
