+++ app/src/main/java/org/gdg/frisbee/android/eventseries/TaggedEventDistanceComparator.java
@@ -1,4 +1,4 @@
-package org.gdg.frisbee.android.utils;
+package org.gdg.frisbee.android.eventseries;
 
 import android.location.Location;
